% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab1.R
\name{mod_tab1_ui}
\alias{mod_tab1_ui}
\alias{mod_tab1_server}
\title{mod_tab1_ui and mod_tab1_server}
\usage{
mod_tab1_ui(id)

mod_tab1_server(input, output, session)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
\details{
tab1 UI Function
}
