% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab6.R
\name{mod_tab6_ui}
\alias{mod_tab6_ui}
\alias{mod_tab6_server}
\title{mod_tab6_ui and mod_tab6_server}
\usage{
mod_tab6_ui(id)

mod_tab6_server(input, output, session)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
\details{
tab6 UI Function
}
